Move to integer money and fix tests.

+++ src/main/kotlin/com/ampnet/crowdfunding/blockchain/contract/EurService.kt
@@ -1,15 +1,15 @@
 package com.ampnet.crowdfunding.blockchain.contract.impl
 
 import org.web3j.crypto.RawTransaction
-import java.math.BigDecimal
+import java.math.BigInteger
 
 interface EurService {
 
-    fun generateMintTransaction(from: String, to: String, amount: BigDecimal): RawTransaction
+    fun generateMintTransaction(from: String, to: String, amount: BigInteger): RawTransaction
 
-    fun generateBurnFromTransaction(from: String, burnFrom: String, amount: BigDecimal): RawTransaction
+    fun generateBurnFromTransaction(from: String, burnFrom: String, amount: BigInteger): RawTransaction
 
-    fun generateApproveTransaction(from: String, approve: String, amount: BigDecimal): RawTransaction
+    fun generateApproveTransaction(from: String, approve: String, amount: BigInteger): RawTransaction
 
-    fun balanceOf(address: String): BigDecimal
+    fun balanceOf(address: String): BigInteger
 }
